+++ tests/src/Command/CoverageCheckCommandTest.php
@@ -15,13 +15,14 @@ declare(strict_types=1);
 
 namespace Esi\CoverageCheck\Tests\Command;
 
+use Esi\CoverageCheck\Application;
 use Esi\CoverageCheck\Command\CoverageCheckCommand;
 use Esi\CoverageCheck\CoverageCheck;
+use Esi\CoverageCheck\Style\CoverageCheckStyle;
 use InvalidArgumentException;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\UsesClass;
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Output\Output;
 use Symfony\Component\Console\Tester\ApplicationTester;
@@ -34,6 +35,8 @@ use function trim;
  * @internal
  */
 #[CoversClass(CoverageCheckCommand::class)]
+#[CoversClass(Application::class)]
+#[CoversClass(CoverageCheckStyle::class)]
 #[UsesClass(CoverageCheck::class)]
 class CoverageCheckCommandTest extends TestCase
 {
@@ -44,16 +47,73 @@ class CoverageCheckCommandTest extends TestCase
     protected function setUp(): void
     {
         self::$fixtures = [
-            'valid'   => dirname(__FILE__, 3) . '/fixtures/clover.xml',
-            'invalid' => dirname(__FILE__, 3) . '/fixtures/clovr.xml',
-            'empty'   => dirname(__FILE__, 3) . '/fixtures/empty.xml',
+            'valid'        => dirname(__FILE__, 3) . '/fixtures/clover.xml',
+            'notexist'     => dirname(__FILE__, 3) . '/fixtures/clovr.xml',
+            'empty'        => dirname(__FILE__, 3) . '/fixtures/empty.xml',
+            'invalid_root' => dirname(__FILE__, 3) . '/fixtures/invalid_root_element.xml',
+            'no_children'  => dirname(__FILE__, 3) . '/fixtures/no_children.xml',
+            'no_metrics'   => dirname(__FILE__, 3) . '/fixtures/no_project_metrics.xml',
         ];
+
+        $command     = new CoverageCheckCommand(new CoverageCheck());
+        $commandName = $command->getName();
 
         $this->application = new Application('PHPUnit Coverage Check', CoverageCheck::VERSION);
         $this->application->setAutoExit(false);
-        $this->application->add(new CoverageCheckCommand(new CoverageCheck()));
-
+        $this->application->add($command);
+        $this->application->setDefaultCommand($commandName, true);
         $this->tester = new ApplicationTester($this->application);
+    }
+
+    public function testCloverFileInvalidRootElement(): void
+    {
+        $this->tester->run([
+            'cloverfile' => self::$fixtures['invalid_root'],
+            'threshold'  => 90,
+        ]);
+
+        self::assertEquals(self::$fixtures['invalid_root'], $this->tester->getInput()->getArgument('cloverfile'));
+        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+
+        self::assertEquals(
+            '[ERROR] Clover file appears to be invalid. Are you sure this is a PHPUnit generated clover report?',
+            trim($this->tester->getDisplay())
+        );
+        self::assertEquals(Command::INVALID, $this->tester->getStatusCode());
+    }
+
+    public function testCloverFileNoChildren(): void
+    {
+        $this->tester->run([
+            'cloverfile' => self::$fixtures['no_children'],
+            'threshold'  => 90,
+        ]);
+
+        self::assertEquals(self::$fixtures['no_children'], $this->tester->getInput()->getArgument('cloverfile'));
+        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+
+        self::assertEquals(
+            '[ERROR] Clover file appears to be invalid. Are you sure this is a PHPUnit generated clover report?',
+            trim($this->tester->getDisplay())
+        );
+        self::assertEquals(Command::INVALID, $this->tester->getStatusCode());
+    }
+
+    public function testCloverFileNoProjectMetrics(): void
+    {
+        $this->tester->run([
+            'cloverfile' => self::$fixtures['no_metrics'],
+            'threshold'  => 90,
+        ]);
+
+        self::assertEquals(self::$fixtures['no_metrics'], $this->tester->getInput()->getArgument('cloverfile'));
+        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+
+        self::assertEquals(
+            '[ERROR] Clover file appears to be invalid. Are you sure this is a PHPUnit generated clover report?',
+            trim($this->tester->getDisplay())
+        );
+        self::assertEquals(Command::INVALID, $this->tester->getStatusCode());
     }
 
     public function testRunInvalidCloverFile(): void
@@ -62,13 +122,8 @@ class CoverageCheckCommandTest extends TestCase
         $this->expectExceptionMessageMatches('/Invalid input file provided. Was given: (.*?)clovr.xml/');
         $tester = new CommandTester($this->application->find('coverage:check'));
         $tester->execute([
-            'command'    => 'coverage:check',
-            'cloverfile' => self::$fixtures['invalid'],
+            'cloverfile' => self::$fixtures['notexist'],
             'threshold'  => 90,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_NORMAL,
         ]);
     }
 
@@ -78,13 +133,8 @@ class CoverageCheckCommandTest extends TestCase
         $this->expectExceptionMessage('The threshold must be a minimum of 1 and a maximum of 100, 101 given');
         $tester = new CommandTester($this->application->find('coverage:check'));
         $tester->execute([
-            'command'    => 'coverage:check',
             'cloverfile' => self::$fixtures['valid'],
             'threshold'  => 101,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_NORMAL,
         ]);
     }
 
@@ -94,33 +144,23 @@ class CoverageCheckCommandTest extends TestCase
         $this->expectExceptionMessage('The threshold must be a minimum of 1 and a maximum of 100, 0 given');
         $tester = new CommandTester($this->application->find('coverage:check'));
         $tester->execute([
-            'command'    => 'coverage:check',
             'cloverfile' => self::$fixtures['valid'],
             'threshold'  => 0,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_NORMAL,
         ]);
     }
 
     public function testRunNotEnoughCode(): void
     {
         $this->tester->run([
-            'command'    => 'coverage:check',
             'cloverfile' => self::$fixtures['empty'],
             'threshold'  => 90,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_VERBOSE,
         ]);
 
         self::assertEquals(self::$fixtures['empty'], $this->tester->getInput()->getArgument('cloverfile'));
         self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
 
         self::assertEquals(
-            'Insufficient data for calculation. Please add more code.',
+            '[ERROR] Insufficient data for calculation. Please add more code.',
             trim($this->tester->getDisplay())
         );
         self::assertEquals(Command::FAILURE, $this->tester->getStatusCode());
@@ -129,22 +169,17 @@ class CoverageCheckCommandTest extends TestCase
     public function testRunNotEnoughCodePercentageOnly(): void
     {
         $this->tester->run([
-            'command'           => 'coverage:check',
             'cloverfile'        => self::$fixtures['empty'],
             'threshold'         => 90,
             '--only-percentage' => true,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_VERBOSE,
-        ]);
+        ], ['verbosity' => Output::VERBOSITY_NORMAL]);
 
         self::assertEquals(self::$fixtures['empty'], $this->tester->getInput()->getArgument('cloverfile'));
         self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
         self::assertTrue($this->tester->getInput()->getOption('only-percentage'));
 
         self::assertEquals(
-            'Insufficient data for calculation. Please add more code.',
+            '[ERROR] Insufficient data for calculation. Please add more code.',
             trim($this->tester->getDisplay())
         );
         self::assertEquals(Command::FAILURE, $this->tester->getStatusCode());
@@ -153,20 +188,15 @@ class CoverageCheckCommandTest extends TestCase
     public function testRunValidNonPassingOptions(): void
     {
         $this->tester->run([
-            'command'    => 'coverage:check',
             'cloverfile' => self::$fixtures['valid'],
             'threshold'  => 100,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_VERBOSE,
         ]);
 
         self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
         self::assertEquals(100, $this->tester->getInput()->getArgument('threshold'));
 
         self::assertEquals(
-            'Total code coverage is 90.32 % which is below the accepted 100 %',
+            '[ERROR] Total code coverage is 90.32% which is below the accepted 100%',
             trim($this->tester->getDisplay())
         );
         self::assertEquals(Command::FAILURE, $this->tester->getStatusCode());
@@ -175,14 +205,9 @@ class CoverageCheckCommandTest extends TestCase
     public function testRunValidOptionsNonPassingPercentageOnly(): void
     {
         $this->tester->run([
-            'command'           => 'coverage:check',
             'cloverfile'        => self::$fixtures['valid'],
             'threshold'         => 100,
             '--only-percentage' => true,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_VERBOSE,
         ]);
 
         self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
@@ -190,7 +215,7 @@ class CoverageCheckCommandTest extends TestCase
         self::assertTrue($this->tester->getInput()->getOption('only-percentage'));
 
         self::assertEquals(
-            '90.32 %',
+            '90.32%',
             trim($this->tester->getDisplay())
         );
         self::assertEquals(Command::FAILURE, $this->tester->getStatusCode());
@@ -199,20 +224,15 @@ class CoverageCheckCommandTest extends TestCase
     public function testRunValidOptionsPassing(): void
     {
         $this->tester->run([
-            'command'    => 'coverage:check',
             'cloverfile' => self::$fixtures['valid'],
             'threshold'  => 90,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_VERBOSE,
         ]);
 
         self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
         self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
 
         self::assertEquals(
-            'Total code coverage is 90.32 % - OK!',
+            '[OK] Total code coverage is 90.32%',
             trim($this->tester->getDisplay())
         );
         self::assertEquals(Command::SUCCESS, $this->tester->getStatusCode());
@@ -221,14 +241,9 @@ class CoverageCheckCommandTest extends TestCase
     public function testRunValidOptionsPassingPercentageOnly(): void
     {
         $this->tester->run([
-            'command'           => 'coverage:check',
             'cloverfile'        => self::$fixtures['valid'],
             'threshold'         => 90,
             '--only-percentage' => true,
-        ], [
-            'interactive' => false,
-            'decorated'   => false,
-            'verbosity'   => Output::VERBOSITY_VERBOSE,
         ]);
 
         self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
@@ -236,7 +251,7 @@ class CoverageCheckCommandTest extends TestCase
         self::assertTrue($this->tester->getInput()->getOption('only-percentage'));
 
         self::assertEquals(
-            '90.32 %',
+            '90.32%',
             trim($this->tester->getDisplay())
         );
         self::assertEquals(Command::SUCCESS, $this->tester->getStatusCode());